menu choice time buffer
it should apply to all the menu choices if I didn't miss anything
you need to press twice to move to another skill after one round
It's not that big of a deal, I just don't know if it has anything to do with my code. It could be this way before my changes, I don't remember.